#1083 fix filter bug
* move reset filter logic to ProjectFilters
* fix bug in useDatePicker
