+++ rdmo/projects/assets/js/components/helper/ProjectFilters.js
@@ -5,7 +5,7 @@ import { get } from 'lodash'
 import DatePicker from 'react-datepicker'
 import { formatISO, set  } from 'date-fns'
 
-import { Select } from 'rdmo/core/assets/js/components'
+import { Link, Select } from 'rdmo/core/assets/js/components'
 import useDatePicker from '../../hooks/useDatePicker'
 import { language } from 'rdmo/core/assets/js/utils'
 
@@ -17,6 +17,24 @@ const ProjectFilters = ({ catalogs, config, configActions, isManager, projectsAc
     setStartDate,
     setEndDate
   } = useDatePicker()
+
+  const showFilters = get(config, 'showFilters', false)
+  const toggleFilters = () => configActions.updateConfig('showFilters', !showFilters)
+
+  const resetAllFilters = () => {
+    configActions.deleteConfig('params.catalog')
+    configActions.deleteConfig('params.created_after')
+    setStartDate('created', null)
+    console.log('dateRange', dateRange)
+    configActions.deleteConfig('params.created_before')
+    setEndDate('created', null)
+    configActions.deleteConfig('params.last_changed_after')
+    setStartDate('last_changed', null)
+    configActions.deleteConfig('params.last_changed_before')
+    setEndDate('last_changed', null)
+    configActions.updateConfig('params.page', '1')
+    projectsActions.fetchProjects()
+  }
 
   const catalogOptions = catalogs?.map(catalog => ({ value: catalog.id.toString(), label: catalog.title }))
   const selectedCatalog = get(config, 'params.catalog', '')
@@ -50,92 +68,106 @@ const ProjectFilters = ({ catalogs, config, configActions, isManager, projectsAc
   }
 
   return (
-    <div className="panel panel-default panel-filters mt-10 mb-0">
-      <div className="panel-body">
-        <div className="row">
-          <div className={`col-md-${isManager ? 4 : 8}`}>
-            <label className="control-label text-muted">{gettext('Filter by catalog')}</label>
-            <div className="search-container">
-              <Select
-                className="select-custom"
-                onChange={updateCatalogFilter}
-                options={catalogOptions ?? []}
-                placeholder={gettext('Select catalog')}
-                value={selectedCatalog}
-              />
-            </div>
-          </div>
-          {isManager && (
-            <div className="col-md-4">
-              <label className="control-label text-muted">{gettext('Filter by created date')}</label>
-              <div className="projects-datepicker">
-                <div className="row">
-                  <div className="col-md-6">
-                    <DatePicker
-                      autoComplete="off"
-                      className="form-control"
-                      dateFormat={dateFormat}
-                      id="created-start-date-picker"
-                      isClearable
-                      locale={getLocale(language)}
-                      onChange={date => handleDateChange('created', 'start', date)}
-                      placeholderText={gettext('Select start date')}
-                      selected={dateRange.createdStart ?? get(config, 'params.created_after', '')}
-                    />
-                  </div>
-                  <div className="col-md-6">
-                    <DatePicker
-                      autoComplete="off"
-                      className="form-control"
-                      dateFormat={dateFormat}
-                      id="created-end-date-picker"
-                      isClearable
-                      locale={getLocale(language)}
-                      onChange={date => handleDateChange('created', 'end', date)}
-                      placeholderText={gettext('Select end date')}
-                      selected={dateRange.createdEnd ?? get(config, 'params.created_before', '')}
-                    />
-                  </div>
+    <>
+      {showFilters && (
+        <div className="panel panel-default panel-filters mt-10 mb-0">
+          <div className="panel-body">
+            <div className="row">
+              <div className={`col-md-${isManager ? 4 : 8}`}>
+                <label className="control-label text-muted">{gettext('Filter by catalog')}</label>
+                <div className="search-container">
+                  <Select
+                    className="select-custom"
+                    onChange={updateCatalogFilter}
+                    options={catalogOptions ?? []}
+                    placeholder={gettext('Select catalog')}
+                    value={selectedCatalog}
+                  />
                 </div>
               </div>
-            </div>
-          )}
-          <div className="col-md-4">
-            <label className="control-label text-muted">{gettext('Filter by last changed date')}</label>
-            <div className="projects-datepicker">
-              <div className="row">
-                <div className="col-md-6">
-                  <DatePicker
-                    autoComplete="off"
-                    className="form-control"
-                    dateFormat={dateFormat}
-                    id="last-changed-start-date-picker"
-                    isClearable
-                    locale={getLocale(language)}
-                    onChange={date => handleDateChange('last_changed', 'start', date)}
-                    placeholderText={gettext('Select start date')}
-                    selected={dateRange.lastChangedStart ?? get(config, 'params.last_changed_after', '')}
-                  />
+              {isManager && (
+                <div className="col-md-4">
+                  <label className="control-label text-muted">{gettext('Filter by created date')}</label>
+                  <div className="projects-datepicker">
+                    <div className="row">
+                      <div className="col-md-6">
+                        <DatePicker
+                          autoComplete="off"
+                          className="form-control"
+                          dateFormat={dateFormat}
+                          id="created-start-date-picker"
+                          isClearable
+                          locale={getLocale(language)}
+                          onChange={date => handleDateChange('created', 'start', date)}
+                          placeholderText={gettext('Select start date')}
+                          selected={dateRange.createdStart ?? get(config, 'params.created_after', '')}
+                        />
+                      </div>
+                      <div className="col-md-6">
+                        <DatePicker
+                          autoComplete="off"
+                          className="form-control"
+                          dateFormat={dateFormat}
+                          id="created-end-date-picker"
+                          isClearable
+                          locale={getLocale(language)}
+                          onChange={date => handleDateChange('created', 'end', date)}
+                          placeholderText={gettext('Select end date')}
+                          selected={dateRange.createdEnd ?? get(config, 'params.created_before', '')}
+                        />
+                      </div>
+                    </div>
+                  </div>
                 </div>
-                <div className="col-md-6">
-                  <DatePicker
-                    autoComplete="off"
-                    className="form-control"
-                    dateFormat={dateFormat}
-                    id="last-changed-end-date-picker"
-                    isClearable
-                    locale={getLocale(language)}
-                    onChange={date => handleDateChange('last_changed', 'end', date)}
-                    placeholderText={gettext('Select end date')}
-                    selected={dateRange.lastChangedEnd ?? get(config, 'params.last_changed_before', '')}
-                  />
+              )}
+              <div className="col-md-4">
+                <label className="control-label text-muted">{gettext('Filter by last changed date')}</label>
+                <div className="projects-datepicker">
+                  <div className="row">
+                    <div className="col-md-6">
+                      <DatePicker
+                        autoComplete="off"
+                        className="form-control"
+                        dateFormat={dateFormat}
+                        id="last-changed-start-date-picker"
+                        isClearable
+                        locale={getLocale(language)}
+                        onChange={date => handleDateChange('last_changed', 'start', date)}
+                        placeholderText={gettext('Select start date')}
+                        selected={dateRange.lastChangedStart ?? get(config, 'params.last_changed_after', '')}
+                      />
+                    </div>
+                    <div className="col-md-6">
+                      <DatePicker
+                        autoComplete="off"
+                        className="form-control"
+                        dateFormat={dateFormat}
+                        id="last-changed-end-date-picker"
+                        isClearable
+                        locale={getLocale(language)}
+                        onChange={date => handleDateChange('last_changed', 'end', date)}
+                        placeholderText={gettext('Select end date')}
+                        selected={dateRange.lastChangedEnd ?? get(config, 'params.last_changed_before', '')}
+                      />
+                    </div>
+                  </div>
                 </div>
               </div>
             </div>
           </div>
         </div>
+      )}
+      <div className="pull-right mt-5">
+        {showFilters && !Object.keys(config.params).every(key => ['ordering', 'page', 'user'].includes(key)) && (
+          <Link className="element-link mr-10 mb-10" onClick={resetAllFilters}>
+            {gettext('Reset all filters')}
+          </Link>
+        )}
+        <Link className="element-link mb-10" onClick={toggleFilters}>
+          {showFilters ? gettext('Hide filters') : gettext('Show filters')}
+        </Link>
       </div>
-    </div>
+    </>
   )
 }
